+++ libs/error-stack/tests/test_frame.rs
@@ -9,8 +9,8 @@ use common::*;
 #[test]
 fn opaque_attachment() {
     let mut report = create_report()
-        .attach(AttachmentA(10))
-        .attach(AttachmentB(20));
+        .attach_opaque(AttachmentA(10))
+        .attach_opaque(AttachmentB(20));
 
     let frame = report.frames_mut().next().expect("No frame found");
     let source = frame
@@ -32,10 +32,10 @@ fn opaque_attachment() {
 
 #[test]
 fn sources() {
-    let mut report_a = create_report().attach(AttachmentA(10)).expand();
-    let report_b = create_report().attach(AttachmentA(20));
+    let mut report_a = create_report().attach_opaque(AttachmentA(10)).expand();
+    let report_b = create_report().attach_opaque(AttachmentA(20));
     report_a.push(report_b);
-    let mut report = report_a.attach(AttachmentB(30));
+    let mut report = report_a.attach_opaque(AttachmentB(30));
 
     let frame = report.frames_mut().next().expect("No frames");
     assert_eq!(frame.sources().len(), 2);
@@ -58,8 +58,8 @@ fn sources() {
 #[test]
 fn printable_attachment() {
     let mut report = create_report()
-        .attach_printable(PrintableA(10))
-        .attach_printable(PrintableB(20));
+        .attach(PrintableA(10))
+        .attach(PrintableB(20));
 
     let frame = report.frames_mut().next().expect("No frame found");
     let source = frame
@@ -109,7 +109,7 @@ fn context() {
 
 #[test]
 fn type_id() {
-    let report = create_report().attach(2_u32);
+    let report = create_report().attach_opaque(2_u32);
     let current = report.current_frame();
 
     assert_eq!(current.type_id(), TypeId::of::<u32>());