+++ libs/error-stack/CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to `error-stack` will be documented in this file.
 
 ### Breaking Changes
 
+- Rename `attach[_lazy]` to `attach_opaque[_with]` and `attach_printable[_lazy]` to `attach[_with]`. ([#7753](https://github.com/hashintel/hash/pull/7753))
 - Set the MSRV to 1.83 ([#5333](https://github.com/hashintel/hash/pull/5333))
 - `Extend` is no longer implemented by `Report<C>`, instead it is implemented on `Report<[C]>`, either use `From` or `Report::expand` to convert between `Report<C>` into `Report<[C]>`. ([#5047](https://github.com/hashintel/hash/pull/5047))
 - `extend_one` has been renamed to `push` and is only implemented on `Report<[C]>`. ([#5047](https://github.com/hashintel/hash/pull/5047))